seek_ecs_tilemap
Module tiles

A module which contains tile components.

Structs

TileBundle

Bundle for tile entities.

TileColor

A custom color for the tile.

TileFlip

Flips the tiles texture along the X, Y or diagonal axes

TilePos

A tile position in the tilemap grid.

TilePosOld

TileStorage

Used to store tile entities for fast look up. Tile entities are stored in a grid. The grid is always filled with None.

TileTextureIndex

A texture index into the atlas or texture array for a single tile. Indices in an atlas are horizontal based.

TileVisible

Hides or shows a tile based on the boolean. Default: True