Module seek_ecs_tilemap::map

A module which contains tilemap components.

Structs

TilemapBundle

TilemapChunks

TilemapGridSize

Size of the tiles on the grid in pixels. This can be used to overlay tiles on top of each other. Ex. A 16x16 pixel tile can be overlapped by 8 pixels by using a grid size of 16x8.

TilemapId

A component which stores a reference to the tilemap entity.

TilemapSize

Size of the tilemap in tiles.

TilemapSpacing

Spacing between tiles in pixels inside of the texture atlas. Defaults to 0.0

TilemapTextureSize

Size of the atlas texture in pixels.

TilemapTileSize

Size of the tiles in pixels

Enums

TilemapType

TilesetTexture